Refactor CdpIamClient unit tests (and prep the integration tests)
